requirements.txt
amqp==5.3.1
asgiref==3.8.1
beautifulsoup4==4.12.3
billiard==4.2.1
celery==5.4.0
certifi==2024.7.4
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
cron-descriptor==1.4.5
dj-database-url==2.3.0
Django==5.1.6
django-celery-beat==2.7.0
django-cors-headers==4.4.0
django-crontab==0.7.1
django-timezone-field==7.1
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
frozendict==2.4.4
gunicorn==23.0.0
h11==0.14.0
html5lib==1.1
idna==3.7
inflection==0.5.1
iniconfig==2.0.0
kombu==5.4.2
lxml==5.2.2
multitasking==0.0.11
numpy==2.0.0
packaging==24.1
pandas==2.2.2
peewee==3.17.6
platformdirs==4.2.2
pluggy==1.5.0
prompt_toolkit==3.0.50
psycopg2==2.9.9
psycopg2-binary==2.9.9
PyJWT==2.8.0
pytest==8.3.2
pytest-django==4.8.0
python-crontab==3.2.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
PyYAML==6.0.2
requests==2.32.3
six==1.16.0
soupsieve==2.5
sqlparse==0.5.0
typing_extensions==4.12.2
tzdata==2024.1
uritemplate==4.1.1
urllib3==2.2.2
vine==5.1.0
wcwidth==0.2.13
webencodings==0.5.1
whitenoise==6.9.0
yfinance==0.2.41
